# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehouse Management API",
        "description": "API for warehouse management system including inventory, stock movements, attendance, and reporting.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{host}",
            "description": "API Server",
            "variables": {
                "host": {
                    "default": "http://localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/api/barang-keluar/export/excel": {
            "get": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "Export barang keluar to Excel",
                "operationId": "5168a9ef1ea432308bbb9bc6ff41f457",
                "responses": {
                    "200": {
                        "description": "Excel file download"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-keluar/{barang_keluar}/print-surat-jalan": {
            "get": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "Print surat jalan for barang keluar",
                "operationId": "be9f4f0dd0e551ae936beab552cb53ec",
                "parameters": [
                    {
                        "name": "barang_keluar",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF surat jalan"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-keluar": {
            "get": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "List all barang keluar",
                "operationId": "9317590bf89efdbf35f202ce729f3aa2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of barang keluar"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "Create barang keluar",
                "operationId": "e17a6150a24e658e935dfb652655ea62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBarangKeluarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Barang keluar created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/barang-keluar/{barang_keluar}": {
            "get": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "Get barang keluar by ID",
                "operationId": "ba287184893d0ac022988d2d5e689883",
                "parameters": [
                    {
                        "name": "barang_keluar",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang keluar detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "Update barang keluar",
                "operationId": "27e9dc1584a32eb1186871a96047a090",
                "parameters": [
                    {
                        "name": "barang_keluar",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBarangKeluarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Barang keluar updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barang Keluar"
                ],
                "summary": "Delete barang keluar",
                "operationId": "edaf7c4744ba8ca155bf08c4d247d3c0",
                "parameters": [
                    {
                        "name": "barang_keluar",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Barang keluar deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreBarangKeluarRequest": {
                "required": [
                    "no_referensi",
                    "gudang_id",
                    "customer_id",
                    "tanggal"
                ],
                "properties": {
                    "no_referensi": {
                        "type": "string",
                        "example": "BK001"
                    },
                    "nomor_surat_jalan": {
                        "type": "string",
                        "nullable": true
                    },
                    "gudang_id": {
                        "type": "integer"
                    },
                    "customer_id": {
                        "type": "integer"
                    },
                    "tanggal": {
                        "type": "string",
                        "format": "date"
                    },
                    "keterangan": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "default": "pending",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected",
                            "delivered",
                            "partial"
                        ]
                    },
                    "dokumen": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Passport",
                "description": "Enter Bearer token obtained from login endpoint"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}